services/auth/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
bcrypt==4.2.1
PyJWT==2.9.0
python-multipart==0.0.19
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0